Handle instances where row contains start and end date
Where the start or end date of a member differs from the start and end of a term, the alternative dates are listed in the row. This commit handles any occurrence where both an alternative start and end are given. (Previously, we could only handle occurences of a single alternative date and the scraper would crash if both dates were listed.)